+++ latest9.9/hej/net/sf/saxon/expr/LookupAllExpression.java
@@ -153,9 +153,10 @@ public class LookupAllExpression extends UnaryExpression {
      * and we assume that a sequence has length 5. The resulting estimates may be used, for
      * example, to reorder the predicates in a filter expression so cheaper predicates are
      * evaluated first.
+     * @return
      */
     @Override
-    public int getCost() {
+    public double getCost() {
         return getBaseExpression().getCost() + 1;
     }
